refactor: move logger utilities into dedicated folder
* Relocated logUtils.ts to new logger/ directory for better organization
* Updated all imports across the codebase to reflect new logger path
* No functional changes
