+++ coauthor-for-vs-code/src/commands/executeCommand.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { ensureArray, getConfig } from '../utils/commonUtils';
-import { debug, initializeLogging } from '../utils/logUtils';
+import { debug, initializeLogging } from '../logger/logUtils';
 
 const CHANNEL = 'ExecuteCommand';
 initializeLogging(CHANNEL);
