+++ coauthor-for-vs-code/src/commands/mergeCommands.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { getConfig } from '../utils/commonUtils';
-import { debug } from '../utils/logUtils';
+import { debug } from '../logger/logUtils';
 
 const CHANNEL = 'MergeCommands';
