+++ coauthor-for-vs-code/src/commands/testCommands.ts
@@ -3,7 +3,7 @@ import {
   bestConnectionMethod,
   bestConnectionMethodAnthropic,
 } from '../latex/textConnection';
-import { info, debug, error, initializeLogging } from '../utils/logUtils';
+import { info, debug, error, initializeLogging } from '../logger/logUtils';
 
 const CHANNEL = 'TestCommands';
 initializeLogging(CHANNEL);
