+++ coauthor-for-vs-code/src/logger/logUtils.ts
