+++ coauthor-for-vs-code/src/webview/messageHandler.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import { workspace } from 'vscode';
-import { debug, info, warn, error } from '../utils/logUtils';
+import { debug, info, warn, error } from '../logger/logUtils';
 import { getWorkspacePath, getRelativePath } from '../utils/fileUtils';
 import {
   listInputFiles,
